requirements.txt
google-genai>=1.0.0
anthropic>=0.40.0
supabase>=2.0.0
pydantic>=2.0.0
structlog>=24.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
fastapi>=0.115.0
uvicorn>=0.30.0
